+++ packages/ice/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v3.0.1
 
 - [fix] lock version of `@ice/bundles`, `@ice/webpack-config` and `@ice/route-manifest`
+- [feat] support external config for server
+- [feat] redirect imports for data loader
 
 ## v3.0.0