+++ packages/webpack-config/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [fix] update version `@swc/core`, keep the same version of `@ice/bundles`
 - [fix] lock version of `@ice/bundles`
+- [feat] support redirect imports
 
 ## 1.0.0